requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
pygame
google-generativeai
SpeechRecognition
gTTS
pillow
